requirements.txt
sounddevice
scipy
sarvamai
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sqlalchemy>=2.0
psycopg2-binary>=2.9
